Move the renderTask to the TaskController
